% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomResponses.R
\name{.CustomResponses}
\alias{.CustomResponses}
\title{Generate Custom Multiple Choice Item Responses Based on Item Scores}
\usage{
.CustomResponses(item_data, n_cat = 4, correct)
}
\arguments{
\item{item_data}{List. Each element is a numeric exam score (between 0 and 1) for a student.}

\item{correct}{Character vector with length of n_item. Specifies the solution to each exam question.}

\item{resp.cat}{Numeric value. The number of response categories per item in the exam.}
}
\value{
A data.frame containing the capital letter responses for as many exams as
specified in GenerateData, where the number of correct responses approximates the respective ability..
}
\description{
For each item score provided in \code{item_data}, this function assigns a capital letter as
multiple choice response. A score of 1 is always transformed into the corresponding letter
in the \code{correct} vector, whereas a score of 0 is transformed into a random other letter within
the range defined in \code{resp.cat}.
}
\examples{
# Suppose a student has an exam score of 0.7, there are 10 items:
item_data = c(1,0,1,0,1,1,1,0,1,1)
# Generate the correct solutions to the 10 exam items:
correct <- sample(LETTERS[1:6], 10, replace = T)
# Generate the response data of that person:
.CustomResponses(item_data, n_cat = 6, correct = correct)

}
