% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateExamData.R
\name{.GenerateExamData}
\alias{.GenerateExamData}
\title{Generate Exam Scores with Controlled Variation Around Ability}
\usage{
.GenerateExamData(ability, n_exam)
}
\arguments{
\item{ability}{Numeric value. A student's latent ability score between 0 and 1.}

\item{n_exam}{Numeric value. Number of exams to generate data for.}
}
\value{
A numeric vector of length \code{n_exam} containing simulated exam scores.
}
\description{
This function simulates exam scores by introducing controlled variance around
a predefined student ability level.
}
\details{
The process consists of three steps:
\enumerate{
\item Sample raw deviation values from a beta distribution,
where the shape parameters control the spread of variance.
\item Center the sampled deviations around the student's ability, ensuring
that the final scores reflect realistic performance fluctuations.
\item Constrain the output scores between 0 and 1, maintaining valid probability-like values.
}

The form of the beta distribution to draw deviation values from depends on the student's ability:  '
\code{shape1} is predefined to 10, whereas \code{shape2} is informed by the students' ability,
assigning a higher value to students with higher ability.
This ensures that higher-performing students have less fluctuation in their exam
scores, while lower-performing students experience more variation.
}
\examples{
# Generate exam scores for a student with ability 0.6 across 5 exams
.GenerateExamData(ability = 0.6, n_exam = 5)
}
