% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateItemData.R
\name{.GenerateItemData}
\alias{.GenerateItemData}
\title{Generate Item Scores Based on Exam Scores}
\usage{
.GenerateItemData(exam_data, n_item)
}
\arguments{
\item{exam_data}{List. Each element is a numeric exam score (between 0 and 1) for a student.}

\item{n_item}{Numeric value. Number of items per exam.}
}
\value{
A numeric vector containing binary item scores (0/1) for a single exam, where the
number of 1's approximates the respective exam score. In the context of the package,
this output is used to construct a matrix with student responses.
}
\description{
For each exam score provided in \code{exam_data}, this function computes the required
number of correct responses by multiplying the exam score by the total number of items
(\code{n_item}) and rounding the result. It then creates a binary response vector of length
\code{n_item}, randomly assigning the computed number of correct responses (coded as 1) and the
remainder as incorrect (coded as 0).
}
\examples{
# Suppose a student has an exam score of 0.7 and there are 10 items:
exam_score <- 0.7
.GenerateItemData(exam_data = list(exam_score), n_item = 10)
}
