% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateResponseData.R
\name{.GenerateResponseData}
\alias{.GenerateResponseData}
\title{Generate Multiple Choice Item Responses Based on Item Scores}
\usage{
.GenerateResponseData(item_data)
}
\arguments{
\item{exam_data}{List. Each element is a numeric exam score (between 0 and 1) for a student.}

\item{n_item}{Numeric value. Number of items per exam.}
}
\value{
A numeric vector containing binary item scores (0/1) for a single exam, where the
number of 1's approximates the respective exam score. In the context of the package,
this output is used to construct a matrix with student responses.
}
\description{
For each item score provided in \code{item_data}, this function assigns a multple choice
response between A and D. A score of 1 is always transformed into A, whereas
a score of 0 is transformed into B, C or D by random.
}
\examples{
# Suppose a student has an exam score of 0.7 and there are 10 items:
.GenerateResponseData(item_data = c(1,0,1,0,1,1,1,0,1,1))

}
